+++ package.json
@@ -22,7 +22,6 @@
     "@tinymce/tinymce-react": "^4.3.0",
     "@wordpress/components": "^25.0.0",
     "@wordpress/core-data": "^6.11.0",
-    "fs-extra": "^11.1.1",
     "react-cropper": "^2.3.3",
     "tinymce": "^6.4.2"
   }
