+++ src/components/modal/style.scss
