+++ src/components/style.scss
@@ -1,0 +1,1 @@
+@import "./modal/style.scss"
