+++ src/main.scss
@@ -1,1 +1,1 @@
-@import "./modal/style.scss"
+@import "./components/style.scss"
